Files for plotting CRYSTAL17 band structures & DOS
This is the latest version of the DOS and BANDS plotting scripts that I've been updating these past few weeks. I made a few important changes, namely:
- Spin-up and spin-down band structures and DOS are now plotted. In DOS plots, the spin-down is to the left and spin-up is to the right. This is useful for analyzing magnetic systems.
- The scripts now support calculations with respect to vacuum. It automatically recognizes if "POTC.POTC.dat" and "POTC.out" files are in the same folder as the BANDS/DOS output files, and does everything automatically. When plotting wrt vacuum, zero represents the vacuum level. Otherwise, it represents the fermi level (arbitrary). Note: for POTC you need to add ghost atom layers to improve the electrostatic potential in vacuum. Also, if your slab is not symmetric, you might need to change what side of the slab you are analyzing. (If you have questions about this let me know)
- Projections are now lines instead of fills, making it easier to visualize if there are several projections in the same plot.
- Colors for projections are now highly distinguishable so there is no confusion between the line colors.

Note: I haven't generalized the BANDS script in terms of k-paths. This is mainly because, in my experience, the paths you want to show are very system dependent. So this will depend on how you set up your BAND.d3 file. It's easy to change the labels in the script though. (check lines 137 to 143 in ipBANDS.py)

Please let me know if there are any improvements or changes you'd like me to make. 

- Danny