Automatic plotting script for band structures
Plotting script for electronic band structures that automatically detects paths from the d3 whether they are in CRYSTAL letter format or seekpath format (numbers with labels on the side). If path is not found, it will prompt the user to define a path manually. 

Note: For generality, the graphs have a large x-axis to accommodate low-symmetry structures that have many paths since labels might overlap. This can be changed to a smaller figure size, depending on the system symmetry, and is left for the user to modify if required (line 439 in the script).